Gestion des la precision decimals sur les prix unitaires des produits.

+++ mysql/tables/llx_product_price.sql
@@ -26,11 +26,11 @@ create table llx_product_price
   tms             timestamp,
   fk_product      integer NOT NULL,
   date_price      datetime NOT NULL,
-  price_level	    tinyint(4) NULL DEFAULT 1,
-  price           double,
-  price_ttc       float(12,4) DEFAULT 0,
+  price_level	  tinyint(4) NULL DEFAULT 1,
+  price           double(16,8),
+  price_ttc       double(16,8) DEFAULT 0,
   price_base_type varchar(3)  DEFAULT 'HT',
-  tva_tx          double NOT NULL,
+  tva_tx          double(6,3)  NOT NULL,
   fk_user_author  integer,
   envente         tinyint DEFAULT 1
 )type=innodb;
